+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Unreleased
+- Fixed `Config Defaults: Compare Config File...` on VS Code 1.40+.
+- Fixed errors when using `Config Defaults: Compare Config File...` and the
+	default file exists but the user file does not.
+
 ## 1.2.0
 - Added `Config Defaults: Compare Config File...` command to compare current
 	configuration with defaults.